+++ Projects/Android/jni/rtcw/src/ui/ui_main.c
@@ -187,7 +187,7 @@ qboolean _UI_IsFullscreen( void );
 #if defined( __MACOS__ )
 #pragma export on
 #endif
-int vmMain( int command, int arg0, int arg1, int arg2, int arg3, int arg4, int arg5, int arg6, int arg7, int arg8, int arg9, int arg10, int arg11  ) {
+intptr_t vmMain( int command, int arg0, int arg1, int arg2, int arg3, int arg4, int arg5, int arg6, int arg7, int arg8, int arg9, int arg10, int arg11  ) {
 #if defined( __MACOS__ )
 #pragma export off
 #endif
